requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-gis==1.0
gunicorn==22.0.0
haversine==2.8.1
numpy==2.0.1
packaging==24.1
psycopg==3.2.1
psycopg-binary==3.2.1
sqlparse==0.5.1
typing_extensions==4.12.2
